Arrays solution until 95

+++ src/arraySolutions60to120/Solution069.java
@@ -16,7 +16,7 @@ public class Solution069 {
         double[] arrayA = new double[n];
         System.out.println("Enter elements of Array A: ");
         for (int i = 0; i < n; i++)
-            arrayA[i] = scanner.nextInt();
+            arrayA[i] = scanner.nextDouble();
 
         scanner.close();
